+++ pycaret/internal/pycaret_experiment/time_series_experiment.py
@@ -1,13 +1,19 @@
-from copy import deepcopy
 import os
-import re
-import math
 
 from sktime.forecasting.model_selection import (
     ExpandingWindowSplitter,
     SlidingWindowSplitter,
 )
 
+from pycaret.utils import _coerce_empty_dataframe_to_none
+from pycaret.utils.time_series import get_sp_from_str
+from pycaret.utils.time_series.forecasting import get_predictions_with_intervals
+
+from pycaret.utils.time_series.forecasting.model_selection import (
+    cross_validate,
+    ForecastingRandomizedSearchCV,
+    ForecastingGridSearchCV,
+)
 
 from pycaret.internal.pycaret_experiment.utils import highlight_setup, MLUsecase
 from pycaret.internal.pycaret_experiment.supervised_experiment import (
@@ -17,23 +23,26 @@ from pycaret.internal.pipeline import (
     estimator_pipeline,
     get_pipeline_fit_kwargs,
 )
+
+from pycaret.utils.time_series import SeasonalPeriod, TSModelTypes
+
 from pycaret.internal.utils import (
     color_df,
-    SeasonalPeriod,
-    TSModelTypes,
-    get_function_params,
+    # get_function_params,
     deep_clone,
 )
 import pycaret.internal.patches.sklearn
-import pycaret.internal.patches.yellowbrick
+
 from pycaret.internal.logging import get_logger
 from pycaret.internal.Display import Display
 
 from pycaret.internal.distributions import *
-from pycaret.internal.validation import *
+
 from pycaret.internal.tunable import TunableMixin
 
 from pycaret.utils.datetime import coerce_datetime_to_period_index
+
+from pycaret.utils.time_series.forecasting import update_additional_scorer_kwargs
 
 import pycaret.containers.metrics.time_series
 import pycaret.containers.models.time_series
@@ -50,22 +59,16 @@ from typing import List, Tuple, Any, Union, Optional, Dict, Generator
 import warnings
 from IPython.utils import io
 import traceback
-import plotly.express as px  # type: ignore
-import plotly.graph_objects as go  # type: ignore
+
 import logging
-from sklearn.base import clone  # type: ignore
-from sklearn.model_selection._validation import _aggregate_score_dicts  # type: ignore
-from sklearn.model_selection import check_cv, ParameterGrid, ParameterSampler  # type: ignore
-from sklearn.model_selection._search import _check_param_grid  # type: ignore
-from sklearn.metrics._scorer import get_scorer, _PredictScorer  # type: ignore
-from collections import defaultdict
-from functools import partial
-from scipy.stats import rankdata  # type: ignore
-from joblib import Parallel, delayed  # type: ignore
+
+# from sklearn.base import clone  # type: ignore
+
+# from sklearn.model_selection._search import _check_param_grid  # type: ignore
 
 
 from sktime.forecasting.base import ForecastingHorizon
-from sktime.utils.validation.forecasting import check_y_X  # type: ignore
+
 from sktime.forecasting.model_selection import SlidingWindowSplitter  # type: ignore
 
 from pycaret.internal.tests.time_series import run_test
@@ -74,602 +77,6 @@ from pycaret.internal.plots.time_series import _plot
 
 warnings.filterwarnings("ignore")
 LOGGER = get_logger()
-
-
-def get_folds(cv, y) -> Generator[Tuple[pd.Series, pd.Series], None, None]:
-    """
-    Returns the train and test indices for the time series data
-    """
-    # https://github.com/alan-turing-institute/sktime/blob/main/examples/window_splitters.ipynb
-    for train_indices, test_indices in cv.split(y):
-        # print(f"Train Indices: {train_indices}, Test Indices: {test_indices}")
-        yield train_indices, test_indices
-
-
-def cross_validate_ts(
-    forecaster,
-    y: pd.Series,
-    X: Optional[Union[pd.Series, pd.DataFrame]],
-    cv,
-    scoring: Dict[str, Union[str, _PredictScorer]],
-    fit_params,
-    n_jobs,
-    return_train_score,
-    error_score=0,
-    verbose: int = 0,
-    **additional_scorer_kwargs,
-) -> Dict[str, np.array]:
-    """Performs Cross Validation on time series data
-
-    Parallelization is based on `sklearn` cross_validate function [1]
-    Ref:
-    [1] https://github.com/scikit-learn/scikit-learn/blob/0.24.1/sklearn/model_selection/_validation.py#L246
-
-
-    Parameters
-    ----------
-    forecaster : [type]
-        Time Series Forecaster that is compatible with sktime
-    y : pd.Series
-        The variable of interest for forecasting
-    X : Optional[Union[pd.Series, pd.DataFrame]]
-        Exogenous Variables
-    cv : [type]
-        [description]
-    scoring : Dict[str, Union[str, _PredictScorer]]
-        Scoring Dictionary. Values can be valid strings that can be converted to
-        callable metrics or the callable metrics directly
-    fit_params : [type]
-        Fit parameters to be used when training
-    n_jobs : [type]
-        Number of cores to use to parallelize. Refer to sklearn for details
-    return_train_score : [type]
-        Should the training scores be returned. Unused for now.
-    error_score : int, optional
-        Unused for now, by default 0
-    verbose : int
-        Sets the verbosity level. Unused for now
-    additional_scorer_kwargs: Dict[str, Any]
-        Additional scorer kwargs such as {`sp`:12} required by metrics like MASE
-
-    Returns
-    -------
-    [type]
-        [description]
-
-    Raises
-    ------
-    Error
-        If fit and score raises any exceptions
-    """
-    try:
-        # # For Debug
-        # n_jobs = 1
-        scoring = _get_metrics_dict_ts(scoring)
-        parallel = Parallel(n_jobs=n_jobs)
-
-        out = parallel(
-            delayed(_fit_and_score)(
-                forecaster=clone(forecaster),
-                y=y,
-                X=X,
-                scoring=scoring,
-                train=train,
-                test=test,
-                parameters=None,
-                fit_params=fit_params,
-                return_train_score=return_train_score,
-                error_score=error_score,
-                **additional_scorer_kwargs,
-            )
-            for train, test in get_folds(cv, y)
-        )
-    # raise key exceptions
-    except Exception:
-        raise
-
-    # Similar to parts of _format_results in BaseGridSearch
-    (test_scores_dict, fit_time, score_time, cutoffs) = zip(*out)
-    test_scores = _aggregate_score_dicts(test_scores_dict)
-
-    return test_scores, cutoffs
-
-
-def _get_metrics_dict_ts(
-    metrics_dict: Dict[str, Union[str, _PredictScorer]]
-) -> Dict[str, _PredictScorer]:
-    """Returns a metrics dictionary in which all values are callables
-    of type _PredictScorer
-
-    Parameters
-    ----------
-    metrics_dict : A metrics dictionary in which some values can be strings.
-        If the value is a string, the corresponding callable metric is returned
-        e.g. Dictionary Value of 'neg_mean_absolute_error' will return
-        make_scorer(mean_absolute_error, greater_is_better=False)
-    """
-    return_metrics_dict = {}
-    for k, v in metrics_dict.items():
-        if isinstance(v, str):
-            return_metrics_dict[k] = get_scorer(v)
-        else:
-            return_metrics_dict[k] = v
-    return return_metrics_dict
-
-
-def _fit_and_score(
-    forecaster,
-    y: pd.Series,
-    X: Optional[Union[pd.Series, pd.DataFrame]],
-    scoring: Dict[str, Union[str, _PredictScorer]],
-    train: np.ndarray,
-    test: np.ndarray,
-    parameters,
-    fit_params,
-    return_train_score,
-    error_score=0,
-    **additional_scorer_kwargs,
-):
-    """Fits the forecaster on a single train split and scores on the test split
-    Similar to _fit_and_score from `sklearn` [1] (and to some extent `sktime` [2]).
-    Difference is that [1] operates on a single fold only, whereas [2] operates on all cv folds.
-    Ref:
-    [1] https://github.com/scikit-learn/scikit-learn/blob/0.24.1/sklearn/model_selection/_validation.py#L449
-    [2] https://github.com/alan-turing-institute/sktime/blob/v0.5.3/sktime/forecasting/model_selection/_tune.py#L95
-
-    Parameters
-    ----------
-    forecaster : [type]
-        Time Series Forecaster that is compatible with sktime
-    y : pd.Series
-        The variable of interest for forecasting
-    X : Optional[Union[pd.Series, pd.DataFrame]]
-        Exogenous Variables
-    scoring : Dict[str, Union[str, _PredictScorer]]
-        Scoring Dictionary. Values can be valid strings that can be converted to
-        callable metrics or the callable metrics directly
-    train : np.ndarray
-        Indices of training samples (integer based indexing).
-    test : np.ndarray
-        Indices of test samples (integer based indexing).
-    parameters : [type]
-        Parameter to set for the forecaster
-    fit_params : [type]
-        Fit parameters to be used when training
-    return_train_score : [type]
-        Should the training scores be returned. Unused for now.
-    error_score : int, optional
-        Unused for now, by default 0
-    **additional_scorer_kwargs: Dict[str, Any]
-            Additional scorer kwargs such as {`sp`:12} required by metrics like MASE
-
-    Raises
-    ------
-    ValueError
-        When test indices do not match predicted indices. This is only for
-        for internal checks and should not be raised when used by external users
-    """
-    if parameters is not None:
-        forecaster.set_params(**parameters)
-
-    y_train, y_test = y[train], y[test]
-    X_train = None if X is None else X.iloc[train]
-    X_test = None if X is None else X.iloc[test]
-
-    #### Fit the forecaster ----
-    start = time.time()
-    try:
-        forecaster.fit(y_train, X_train, **fit_params)
-    except Exception as error:
-        logging.error(f"Fit failed on {forecaster}")
-        logging.error(error)
-
-        if error_score == "raise":
-            raise
-
-    fit_time = time.time() - start
-
-    #### Determine Cutoff ----
-    # NOTE: Cutoff is available irrespective of whether fit passed or failed
-    cutoff = forecaster.cutoff
-
-    #### Score the model ----
-    lower = pd.Series([])
-    upper = pd.Series([])
-    if forecaster.is_fitted:
-        # TODO: Add alpha here???
-        y_pred, lower, upper = get_predictions_with_intervals(
-            forecaster=forecaster, X=X_test
-        )
-
-        if (y_test.index.values != y_pred.index.values).any():
-            print(
-                f"\t y_train: {y_train.index.values},"
-                f"\n\t y_test: {y_test.index.values}"
-            )
-            print(f"\t y_pred: {y_pred.index.values}")
-            raise ValueError(
-                "y_test indices do not match y_pred_indices or split/prediction "
-                "length does not match forecast horizon."
-            )
-
-    start = time.time()
-    fold_scores = {}
-    scoring = _get_metrics_dict_ts(scoring)
-
-    # SP should be passed from outside in additional_scorer_kwargs already
-    additional_scorer_kwargs = update_additional_scorer_kwargs(
-        initial_kwargs=additional_scorer_kwargs,
-        y_train=y_train,
-        lower=lower,
-        upper=upper,
-    )
-    for scorer_name, scorer in scoring.items():
-        if forecaster.is_fitted:
-            # get all kwargs in additional_scorer_kwargs
-            # that correspond to parameters in function signature
-            kwargs = {
-                **{
-                    k: v
-                    for k, v in additional_scorer_kwargs.items()
-                    if k in get_function_params(scorer._score_func)
-                },
-                **scorer._kwargs,
-            }
-            metric = scorer._score_func(y_true=y_test, y_pred=y_pred, **kwargs)
-        else:
-            metric = None
-        fold_scores[scorer_name] = metric
-    score_time = time.time() - start
-
-    return fold_scores, fit_time, score_time, cutoff
-
-
-class BaseGridSearch:
-    """
-    Parallelization is based predominantly on [1]. Also similar to [2]
-
-    Ref:
-    [1] https://github.com/scikit-learn/scikit-learn/blob/0.24.1/sklearn/model_selection/_search.py#L795
-    [2] https://github.com/scikit-optimize/scikit-optimize/blob/v0.8.1/skopt/searchcv.py#L410
-    """
-
-    def __init__(
-        self,
-        forecaster,
-        cv,
-        n_jobs=None,
-        pre_dispatch=None,
-        refit: bool = False,
-        refit_metric: str = "smape",
-        scoring=None,
-        verbose=0,
-        error_score=None,
-        return_train_score=None,
-    ):
-        self.forecaster = forecaster
-        self.cv = cv
-        self.n_jobs = n_jobs
-        self.pre_dispatch = pre_dispatch
-        self.refit = refit
-        self.refit_metric = refit_metric
-        self.scoring = scoring
-        self.verbose = verbose
-        self.error_score = error_score
-        self.return_train_score = return_train_score
-
-        self.best_params_ = {}
-        self.cv_results_ = {}
-
-    def fit(
-        self,
-        y: pd.Series,
-        X: Optional[pd.DataFrame] = None,
-        additional_scorer_kwargs: Optional[Dict[str, Any]] = None,
-        **fit_params,
-    ) -> "BaseGridSearch":
-        """Run fit with all sets of parameters.
-
-        Parameters
-        ----------
-        y : pd.Series
-            Target
-        X : Optional[pd.DataFrame], optional
-            Exogenous Variables, by default None
-        additional_scorer_kwargs: Dict[str, Any]
-            Additional scorer kwargs such as {`sp`:12} required by metrics like MASE
-        **fit_params: Dict[str, Any]
-            Parameters passed to the ``fit`` method of the estimator
-
-        Returns
-        -------
-        BaseGridSearch
-            Grid Search Object returned to allow chaining
-
-        Raises
-        ------
-        ValueError
-            When any of the following is True
-            (1) Metric can not be found
-            (2) No candidate provided in search
-            (3) CV Iterator is empty
-        """
-        if additional_scorer_kwargs is None:
-            additional_scorer_kwargs = {}
-
-        y, X = check_y_X(y, X)
-
-        # validate cross-validator
-        cv = check_cv(self.cv)
-        base_forecaster = clone(self.forecaster)
-
-        # This checker is sktime specific and only support 1 metric
-        # Removing for now since we can have multiple metrics
-        # TODO: Add back later if it supports multiple metrics
-        # scoring = check_scoring(self.scoring)
-        # Multiple metrics supported
-        scorers = self.scoring  # Dict[str, Union[str, scorer]]  Not metrics container
-        scorers = _get_metrics_dict_ts(scorers)
-        refit_metric = self.refit_metric
-        if refit_metric not in list(scorers.keys()):
-            raise ValueError(
-                f"Refit Metric: '{refit_metric}' is not available. ",
-                f"Available Values are: {list(scorers.keys())}",
-            )
-
-        results = {}
-        all_candidate_params = []
-        all_out = []
-
-        def evaluate_candidates(candidate_params):
-            candidate_params = list(candidate_params)
-            n_candidates = len(candidate_params)
-            n_splits = cv.get_n_splits(y)
-
-            if self.verbose > 0:
-                print(  # noqa
-                    f"Fitting {n_splits} folds for each of {n_candidates} "
-                    f"candidates, totalling {n_candidates * n_splits} fits"
-                )
-
-            parallel = Parallel(
-                n_jobs=self.n_jobs, verbose=self.verbose, pre_dispatch=self.pre_dispatch
-            )
-            out = parallel(
-                delayed(_fit_and_score)(
-                    forecaster=clone(base_forecaster),
-                    y=y,
-                    X=X,
-                    scoring=scorers,
-                    train=train,
-                    test=test,
-                    parameters=parameters,
-                    fit_params=fit_params,
-                    return_train_score=self.return_train_score,
-                    error_score=self.error_score,
-                    **additional_scorer_kwargs,
-                )
-                for parameters in candidate_params
-                for train, test in get_folds(cv, y)
-            )
-
-            if len(out) < 1:
-                raise ValueError(
-                    "No fits were performed. "
-                    "Was the CV iterator empty? "
-                    "Were there no candidates?"
-                )
-
-            all_candidate_params.extend(candidate_params)
-            all_out.extend(out)
-
-            nonlocal results
-            results = self._format_results(
-                all_candidate_params, scorers, all_out, n_splits
-            )
-            return results
-
-        self._run_search(evaluate_candidates)
-
-        self.best_index_ = results["rank_test_%s" % refit_metric].argmin()
-        self.best_score_ = results["mean_test_%s" % refit_metric][self.best_index_]
-        self.best_params_ = results["params"][self.best_index_]
-
-        self.best_forecaster_ = clone(base_forecaster).set_params(**self.best_params_)
-
-        if self.refit:
-            refit_start_time = time.time()
-            self.best_forecaster_.fit(y, X, **fit_params)
-            self.refit_time_ = time.time() - refit_start_time
-
-        # Store the only scorer not as a dict for single metric evaluation
-        self.scorer_ = scorers
-
-        self.cv_results_ = results
-        self.n_splits_ = cv.get_n_splits(y)
-
-        self._is_fitted = True
-        return self
-
-    @staticmethod
-    def _format_results(candidate_params, scorers, out, n_splits):
-        """From sklearn and sktime"""
-        n_candidates = len(candidate_params)
-        (test_scores_dict, fit_time, score_time, cutoffs) = zip(*out)
-        test_scores_dict = _aggregate_score_dicts(test_scores_dict)
-
-        results = {}
-
-        # From sklearn (with the addition of greater_is_better from sktime)
-        # INFO: For some reason, sklearn func does not work with sktime metrics
-        # without passing greater_is_better (as done in sktime) and processing
-        # it as such.
-        def _store(
-            key_name,
-            array,
-            weights=None,
-            splits=False,
-            rank=False,
-            greater_is_better=False,
-        ):
-            """A small helper to store the scores/times to the cv_results_"""
-            # When iterated first by splits, then by parameters
-            # We want `array` to have `n_candidates` rows and `n_splits` cols.
-            array = np.array(array, dtype=np.float64).reshape(n_candidates, n_splits)
-            if splits:
-                for split_idx in range(n_splits):
-                    # Uses closure to alter the results
-                    results["split%d_%s" % (split_idx, key_name)] = array[:, split_idx]
-
-            array_means = np.average(array, axis=1, weights=weights)
-            results["mean_%s" % key_name] = array_means
-
-            if key_name.startswith(("train_", "test_")) and np.any(
-                ~np.isfinite(array_means)
-            ):
-                warnings.warn(
-                    f"One or more of the {key_name.split('_')[0]} scores "
-                    f"are non-finite: {array_means}",
-                    category=UserWarning,
-                )
-
-            # Weighted std is not directly available in numpy
-            array_stds = np.sqrt(
-                np.average(
-                    (array - array_means[:, np.newaxis]) ** 2, axis=1, weights=weights
-                )
-            )
-            results["std_%s" % key_name] = array_stds
-
-            if rank:
-                # This section is taken from sktime
-                array_means = -array_means if greater_is_better else array_means
-                results["rank_%s" % key_name] = np.asarray(
-                    rankdata(array_means, method="min"), dtype=np.int32
-                )
-
-        _store("fit_time", fit_time)
-        _store("score_time", score_time)
-        # Use one MaskedArray and mask all the places where the param is not
-        # applicable for that candidate. Use defaultdict as each candidate may
-        # not contain all the params
-        param_results = defaultdict(
-            partial(
-                np.ma.MaskedArray, np.empty(n_candidates,), mask=True, dtype=object,
-            )
-        )
-        for cand_i, params in enumerate(candidate_params):
-            for name, value in params.items():
-                # An all masked empty array gets created for the key
-                # `"param_%s" % name` at the first occurrence of `name`.
-                # Setting the value at an index also unmasks that index
-                param_results["param_%s" % name][cand_i] = value
-
-        results.update(param_results)
-        # Store a list of param dicts at the key "params"
-        results["params"] = candidate_params
-
-        for scorer_name, scorer in scorers.items():
-            # Computed the (weighted) mean and std for test scores alone
-            _store(
-                "test_%s" % scorer_name,
-                test_scores_dict[scorer_name],
-                splits=True,
-                rank=True,
-                weights=None,
-                greater_is_better=True if scorer._sign == 1 else False,
-            )
-
-        return results
-
-
-class ForecastingGridSearchCV(BaseGridSearch):
-    """Exhaustive search over specified parameter values for an estimator.
-
-    TODO: Add detailed docstring similar to the one available here:
-    https://github.com/scikit-learn/scikit-learn/blob/c6512929fbee7232949c0f18cfb28cf3b5959df9/sklearn/model_selection/_search.py#L972
-    """
-
-    def __init__(
-        self,
-        forecaster,
-        cv,
-        param_grid,
-        scoring=None,
-        n_jobs=None,
-        refit=True,
-        refit_metric: str = "smape",
-        verbose=0,
-        pre_dispatch="2*n_jobs",
-        error_score=np.nan,
-        return_train_score=False,
-    ):
-        super(ForecastingGridSearchCV, self).__init__(
-            forecaster=forecaster,
-            cv=cv,
-            n_jobs=n_jobs,
-            pre_dispatch=pre_dispatch,
-            refit=refit,
-            refit_metric=refit_metric,
-            scoring=scoring,
-            verbose=verbose,
-            error_score=error_score,
-            return_train_score=return_train_score,
-        )
-        self.param_grid = param_grid
-        _check_param_grid(param_grid)
-
-    def _run_search(self, evaluate_candidates):
-        """Search all candidates in param_grid"""
-        evaluate_candidates(ParameterGrid(self.param_grid))
-
-
-class ForecastingRandomizedSearchCV(BaseGridSearch):
-    """Randomized search on hyper parameters.
-
-    TODO: Add detailed docstring similar to the one available here:
-    https://github.com/scikit-learn/scikit-learn/blob/c6512929fbee7232949c0f18cfb28cf3b5959df9/sklearn/model_selection/_search.py#L1292
-    """
-
-    def __init__(
-        self,
-        forecaster,
-        cv,
-        param_distributions,
-        n_iter=10,
-        scoring=None,
-        n_jobs=None,
-        refit=True,
-        refit_metric: str = "smape",
-        verbose=0,
-        random_state=None,
-        pre_dispatch="2*n_jobs",
-        error_score=np.nan,
-        return_train_score=False,
-    ):
-        super(ForecastingRandomizedSearchCV, self).__init__(
-            forecaster=forecaster,
-            cv=cv,
-            n_jobs=n_jobs,
-            pre_dispatch=pre_dispatch,
-            refit=refit,
-            refit_metric=refit_metric,
-            scoring=scoring,
-            verbose=verbose,
-            error_score=error_score,
-            return_train_score=return_train_score,
-        )
-        self.param_distributions = param_distributions
-        self.n_iter = n_iter
-        self.random_state = random_state
-
-    def _run_search(self, evaluate_candidates):
-        """Search n_iter candidates from param_distributions"""
-        return evaluate_candidates(
-            ParameterSampler(
-                self.param_distributions, self.n_iter, random_state=self.random_state
-            )
-        )
 
 
 class TimeSeriesExperiment(_SupervisedExperiment):
@@ -744,45 +151,35 @@ class TimeSeriesExperiment(_SupervisedExperiment):
         functions = pd.DataFrame(
             [
                 ["session_id", self.seed],
-                # ["Target", self.target_param],
+                ["Target", self.target_param],
                 ["Original Data", self.data_before_preprocess.shape],
                 ["Missing Values", kwargs["missing_flag"]],
+                ["Forecasting Type", self.forecasting_type],
+                ["Transformed Train Target", self.y_train.shape],
+                ["Transformed Test Target", self.y_test.shape],
+                ["Transformed Train Exogenous", self.X_train.shape],
+                ["Transformed Test Exogenous", self.X_test.shape],
+                ["Fold Generator", type(self.fold_generator).__name__],
+                ["Fold Number", self.fold_param],
+                ["Enforce Prediction Interval", self.enforce_pi],
+                ["Seasonal Period Tested", self.seasonal_period],
+                ["Seasonality Detected", self.seasonality_present],
+                ["Seasonality Used in Models", self.sp_to_use],
+                ["Target Strictly Positive", self.strictly_positive],
+                ["Target White Noise", self.white_noise],
+                ["Recommended d", self.lowercase_d],
+                ["Recommended Seasonal D", self.uppercase_d],
+                ["CPU Jobs", self.n_jobs_param],
+                ["Use GPU", self.gpu_param],
+                ["Log Experiment", self.logging_param],
+                ["Experiment Name", self.exp_name_log],
+                ["USI", self.USI],
             ]
-            + (
-                [
-                    ["Forecasting Type", self.forecasting_type],
-                    ["Transformed Train Target", self.y_train.shape],
-                    ["Transformed Test Target", self.y_test.shape],
-                    ["Transformed Train Exogenous", self.X_train.shape],
-                    ["Transformed Test Exogenous", self.X_test.shape],
-                    ["Fold Generator", type(self.fold_generator).__name__],
-                    ["Fold Number", self.fold_param],
-                    ["Enforce Prediction Interval", self.enforce_pi],
-                    ["Seasonal Period Tested", self.seasonal_period],
-                    ["Seasonality Detected", self.seasonality_present],
-                    ["Seasonality Used in Models", self.sp_to_use],
-                    ["Target Strictly Positive", self.strictly_positive],
-                    ["Target White Noise", self.white_noise],
-                    ["Recommended d", self.lowercase_d],
-                    ["Recommended Seasonal D", self.uppercase_d],
-                    ["CPU Jobs", self.n_jobs_param],
-                    ["Use GPU", self.gpu_param],
-                    ["Log Experiment", self.logging_param],
-                    ["Experiment Name", self.exp_name_log],
-                    ["USI", self.USI],
-                ]
-            )
             + (
                 [["Imputation Type", kwargs["imputation_type"]],]
                 if self.preprocess
                 else []
             ),
-            # + (
-            #    [
-            #        ["Transform Target", self.transform_target_param],
-            #        ["Transform Target Method", self.transform_target_method_param],
-            #    ]
-            # ),
             columns=["Description", "Value"],
         )
         return functions.style.apply(highlight_setup)
@@ -1834,7 +1231,7 @@ class TimeSeriesExperiment(_SupervisedExperiment):
         model_fit_start = time.time()
 
         additional_scorer_kwargs = self.get_additional_scorer_kwargs()
-        scores, cutoffs = cross_validate_ts(
+        scores, cutoffs = cross_validate(
             # Commented out since supervised_experiment also does not clone
             # when doing cross_validate
             # forecaster=clone(model),
@@ -4196,159 +3593,4 @@ class TimeSeriesExperiment(_SupervisedExperiment):
         """
         additional_scorer_kwargs = {"sp": self.sp_to_use}
         return additional_scorer_kwargs
-
-
-def get_predictions_with_intervals(
-    forecaster, X: pd.DataFrame, fh=None, alpha: float = 0.05
-) -> Tuple[pd.Series, pd.Series, pd.Series]:
-    """Returns the predictions, lower and upper interval values for a
-    forecaster. If the forecaster does not support prediction intervals,
-    then NAN is returned for lower and upper intervals.
-
-    Parameters
-    ----------
-    forecaster : sktime compatible forecaster
-        Forecaster to be used to get the predictions
-    X : pd.DataFrame
-        Exogenous Variables
-    alpha : float, default = 0.05
-        alpha value for prediction interval
-
-    Returns
-    -------
-    Tuple[pd.Series, pd.Series, pd.Series]
-        Predictions, Lower and Upper Interval Values
-    """
-    # Predict and get lower and upper intervals
-    return_pred_int = forecaster.get_tag("capability:pred_int")
-
-    return_values = forecaster.predict(
-        fh=fh, X=X, return_pred_int=return_pred_int, alpha=alpha
-    )
-
-    if return_pred_int:
-        y_pred = return_values[0]
-        lower = return_values[1]["lower"]
-        upper = return_values[1]["upper"]
-    else:
-        y_pred = return_values
-        lower = pd.Series([np.nan] * len(y_pred))
-        upper = pd.Series([np.nan] * len(y_pred))
-        lower.index = y_pred.index
-        upper.index = y_pred.index
-
-    # Prophet with return_pred_int = True returns datetime index.
-    for series in [y_pred, lower, upper]:
-        if isinstance(series.index, pd.DatetimeIndex):
-            series.index = series.index.to_period()
-
-    return y_pred, lower, upper
-
-
-def update_additional_scorer_kwargs(
-    initial_kwargs: Dict[str, Any],
-    y_train: pd.Series,
-    lower: pd.Series,
-    upper: pd.Series,
-) -> Dict[str, Any]:
-    """Updates the initial kwargs with additional scorer kwargs
-    NOTE: Initial kwargs are obtained from experiment, e.g. {"sp": 12} and
-    are common to all folds.
-    The additional kwargs such as y_train, lower, upper are specific to each
-    fold and must be updated dynamically as such.
-
-    Parameters
-    ----------
-    initial_kwargs : Dict[str, Any]
-        Initial kwargs are obtained from experiment, e.g. {"sp": 12} and
-        are common to all folds
-    y_train : pd.Series
-        Training Data. Used in metrics such as MASE
-    lower : pd.Series
-        Lower Limits of Predictions. Used in metrics such as INPI
-    upper : pd.Series
-        Upper Limits of Predictions. Used in metrics such as INPI
-
-    Returns
-    -------
-    Dict[str, Any]
-        Updated kwargs dictionary
-    """
-    additional_scorer_kwargs = initial_kwargs.copy()
-    additional_scorer_kwargs.update(
-        {"y_train": y_train, "lower": lower, "upper": upper,}
-    )
-    return additional_scorer_kwargs
-
-
-def get_sp_from_str(str_freq: str) -> int:
-    """Takes the seasonal period as string detects if it is alphanumeric and returns its integer equivalent.
-        For example -
-        input - '30W'
-        output - 26
-        explanation - we take the lcm of 30 and 52 ( as W = 52) which in this case is 780.
-        And the output is ( lcm / prefix). Here, 780 / 30 = 26.
-
-    Parameters
-    ----------
-    str_freq : str
-        frequency of the dataset passed as a string
-
-    Returns
-    -------
-    int
-        integer equivalent of the string frequency
-
-    Raises
-    ------
-    ValueError
-        if the frequency suffix does not correspond to any of the values in the class SeasonalPeriod then the error is thrown.
-    """
-    str_freq = str_freq.split("-")[0] or str_freq
-    # Checking whether the index_freq contains both digit and alphabet
-    if bool(re.search(r"\d", str_freq)):
-        temp = re.compile("([0-9]+)([a-zA-Z]+)")
-        res = temp.match(str_freq).groups()
-        # separating the digits and alphabets
-        if res[1] in SeasonalPeriod.__members__:
-            prefix = int(res[0])
-            value = SeasonalPeriod[res[1]].value
-            lcm = abs(value * prefix) // math.gcd(value, prefix)
-            seasonal_period = int(lcm / prefix)
-            return seasonal_period
-        else:
-            raise ValueError(
-                f"Unsupported Period frequency: {str_freq}, valid Period frequency suffixes are: {', '.join(SeasonalPeriod.__members__.keys())}"
-            )
-    else:
-
-        if str_freq in SeasonalPeriod.__members__:
-            seasonal_period = SeasonalPeriod[str_freq].value
-            return seasonal_period
-        else:
-            raise ValueError(
-                f"Unsupported Period frequency: {str_freq}, valid Period frequency suffixes are: {', '.join(SeasonalPeriod.__members__.keys())}"
-            )
-
-
-def _coerce_empty_dataframe_to_none(
-    data: Optional[pd.DataFrame],
-) -> Optional[pd.DataFrame]:
-    """Returns None if the data is an empty dataframe or None,
-    else return the dataframe as is.
-
-    Parameters
-    ----------
-    data : Optional[pd.DataFrame]
-        Dataframe to be checked or None
-
-    Returns
-    -------
-    Optional[pd.DataFrame]
-        Returned Dataframe OR None (if dataframe is empty or None)
-    """
-    if isinstance(data, pd.DataFrame) and data.empty:
-        return None
-    else:
-        return data
